fixed type error

+++ xlearn/models.py
@@ -232,7 +232,7 @@ def transformer3_pooling(ih, iw, nb_conv, size_conv, nb_gpu):
     fc1 = Dropout(0.25)(fc1)
     fc1 = Dense(iw * ih / 64, activation='relu')(fc1)
     fc1 = Dropout(0.25)(fc1)
-    fc1 = Reshape((ih / 8, iw / 8, 1))(fc1)
+    fc1 = Reshape((int(ih / 8), int(iw / 8), 1))(fc1)
 
     fc2 = Conv2DTranspose(nb_conv * 4, (size_conv, size_conv), activation='relu', padding='same')(fc1)
     fc2 = Conv2DTranspose(nb_conv * 8, (size_conv, size_conv), activation='relu', padding='same')(fc2)
